Extract the statistics-plot binning out of PlotModelStat
Last step of phase 1. SpectrumMatchHistograms.GetHistogramData returns the bins for
all 16 statistics plots as plain data, and PlotModelStat now delegates to it. The
point is phase 3: when the OxyPlot rendering of these histograms is rewritten,
"did the numbers move?" becomes a question about data rather than about a chart.

This is a pure move, and that is verified rather than asserted - diffing the
relocated binning against the original region line by line comes back identical.
Only the signature changed (psmsBySourceFile is now a parameter rather than a field
read) plus visibility.

roundToBin and NormalizeAmbiguityLevel moved with it rather than being duplicated;
PlotModelStat calls the shared ones. Both were private statics used on either side
of the split, which is exactly the case where two copies drift.

17 characterisation tests, cross-platform. They deliberately assert invariants
rather than golden numbers: every plot type bins without throwing on real
ambiguity-containing input, each bin dictionary accounts for exactly the values
collected per source file, bin sizes are pinned for the four plots where a
re-binning would be invisible, and pipe-delimited notches are split across bins
rather than parsed as one value. VariantCrossTest.psmtsv is linked into the
cross-platform project as the fixture - 15 rows, and it carries ambiguous notches,
which is the case that breaks a naive double.Parse.

245 tests pass on macOS arm64, up from 228. Solution builds in Release bar WiX.
